% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demo.R
\name{demo}
\alias{demo}
\title{Funcion de demo}
\usage{
demo(x)
}
\arguments{
\item{x}{A \code{integer(1)} or \code{numeric(1)}.}
}
\value{
\code{x^2}
}
\description{
Esta función calcula el cuadrado del valor inicial.
}
\examples{
demo(2)
}
